frontend/requirements.txt
pydantic==1.10.2
finnhub-python==2.4.15
uvicorn==0.20.0
numpy==1.23.2
pandas==1.4.3
boto3==1.26.44
plotly==5.11.0
mlflow==2.0.1
hyperopt==0.2.7
scikit-learn==1.2.0
xgboost==1.7.2
streamlit==1.16.0